+++ CHANGELOG.md
@@ -15,6 +15,19 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - example of modal with table
 
+## [9.83.0] - 2019-10-01
+
+### Fixed
+
+- `InputSearch` update component style with figma's design.
+- `Input` add `onMouseEnter` and `onMouseLeave` events support.
+
+## [9.82.0] - 2019-10-01
+
+### Added
+
+- New component to split elements.
+
 ## [9.81.4] - 2019-09-26
 
 ### Fixed